packages/kagent/pyproject.toml
[project]
name = "kagent"
version = "0.1.0"
description = "Core agent framework for the k agent framework"
readme = "README.md"
authors = [
    { name = "Xiangzhuang Shen", email = "datahonor@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "kai",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = ["kagent"]
source-exclude = ["tests/**/*"]

[tool.uv.sources]
kai = { workspace = true }

[tool.ruff]
line-length = 100

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.12"
reportUnnecessaryTypeIgnoreComment = "warning"
include = [
    "src/**/*.py",
    "tests/**/*.py",
    "examples/**/*.py",
]
